+++ go.mod
@@ -2,4 +2,4 @@ module shopify-adaptor
 
 go 1.20
 
-require github.com/joho/godotenv v1.5.1 // indirect
+require github.com/joho/godotenv v1.5.1
